nanoflann 1.6 knnSearch: use unsigned int index buffers
nanoflann 1.6's KDTreeSingleIndexAdaptor::knnSearch signature takes
IndexType* out_indices, where IndexType defaults to unsigned int.
The two knnSearch callers in this fork previously handed in a
size_t (long unsigned int) buffer, which no longer converts.

Change the `indices` QVectors from size_t to unsigned int in both
neighbor_query methods, and update the PointCloudSearchResult
nearest-neighbor constructor in cnet2dem to accept the matching
type. The downstream `m_pc->point(indices[i])` call continues to
work since unsigned int converts implicitly.

+++ isis/src/control/apps/cnet2dem/PointCloudSearchResult.h
@@ -57,8 +57,9 @@ namespace Isis {
         m_pc            = QSharedPointer<PointCloud<T,D> > (0);
       }
 
+      // nanoflann >=1.6 gives back index buffers as IndexType (unsigned int).
       PointCloudSearchResult(const T &source, int neighbors,
-                      QVector<size_t> &indices, QVector<double> &distances,
+                      QVector<unsigned int> &indices, QVector<double> &distances,
                       QSharedPointer<PointCloud<T,D> > &pc) {
         Q_ASSERT ( indices.size() == distances.size() );
         Q_ASSERT ( neighbors == indices.size() );
