+++ isis/src/control/apps/cnetcombinept/PointCloudTree.h
@@ -139,7 +139,9 @@ template <class T> class PointCloudTree {
      * @return @b QList<T> A list of the neighbors.
      */
     QList<T> neighbor_query(const T &point, const int &neighbors) {
-      QVector<size_t> indices(neighbors);
+      // nanoflann >=1.6 knnSearch expects the index buffer as
+      // IndexType* (unsigned int*) rather than size_t*.
+      QVector<unsigned int> indices(neighbors);
       QVector<double> distances(neighbors);
       m_kd_index.knnSearch(point.array(), neighbors, indices.data(),
                            distances.data());
